backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Google Gemini
google-generativeai>=0.5.0

# LangChain RAG
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.11
chromadb>=0.4.0
pypdf>=3.0.0

# Async HTTP
httpx>=0.24.0
aiofiles>=22.1.0
